Revert "fix: To use complete range for catalog entries index internal used datatype was switched from short to int."
This reverts commit 62521a1c1f840c8f56fbb9377aaff249ef15fea9.

+++ src/DataService/Rest/CatalogEntry.cs
@@ -59,7 +59,7 @@ namespace Zeiss.IMT.PiWeb.Api.DataService.Rest
 		/// Gets or sets the unique key of this catalog entry.
 		/// </summary>
 		[JsonProperty( "key" )]
-		public int Key { get; set; }
+		public short Key { get; set; }
 
 		#endregion
